copy from clang, move multi test
